Optimized the code by helper function - remove extra lines
