+++ src/components/Dashboard/posts/post-image.tsx
@@ -7,7 +7,7 @@ import { Separator } from "@/src/components/ui/separator"
 import PostComments from "./post-comments"
 import PostCommentForm from "./post-comment-form"
 import PostCommentsSection from "./post-comments-section"
-import { useRouter, useParams } from "next/navigation"
+import { usePostNavigation } from "@/src/hooks/usePostNavigation"
 
 type Props = {
   post: SelectFilePost
@@ -15,23 +15,10 @@ type Props = {
 }
 
 const ImagePost: React.FC<Props> = ({ post, spaceId }) => {
-  const router = useRouter()
-  const params = useParams()
+  const { navigateToPost } = usePostNavigation()
 
   const handleContentClick = () => {
-    if (spaceId && spaceId !== "shared") {
-      const channelSlug = params?.channel_slug as string
-      const spaceSlug = params?.space_slug as string
-      if (channelSlug && spaceSlug) {
-        router.push(
-          `/channels/${channelSlug}/spaces/${spaceSlug}?page-type=posts&post-id=${post.id}`
-        )
-      } else {
-        router.push(`/posts/${post.id}`)
-      }
-    } else {
-      router.push(`/posts/${post.id}`)
-    }
+    navigateToPost(post.id, spaceId)
   }
 
   return (
