+++ src/components/Dashboard/posts/post-text.tsx
@@ -6,7 +6,7 @@ import PostCommentForm from "./post-comment-form"
 import { SelectComment, SelectPost } from "@/src/db/schema"
 import { Badge } from "../../ui/badge"
 import PostCommentsSection from "./post-comments-section"
-import { useRouter, useParams } from "next/navigation"
+import { usePostNavigation } from "@/src/hooks/usePostNavigation"
 
 type Props = {
   post: SelectPost
@@ -14,23 +14,10 @@ type Props = {
 }
 
 const TextPost: React.FC<Props> = ({ post, spaceId }) => {
-  const router = useRouter()
-  const params = useParams()
+  const { navigateToPost } = usePostNavigation()
 
   const handleContentClick = () => {
-    if (spaceId && spaceId !== "shared") {
-      const channelSlug = params?.channel_slug as string
-      const spaceSlug = params?.space_slug as string
-      if (channelSlug && spaceSlug) {
-        router.push(
-          `/channels/${channelSlug}/spaces/${spaceSlug}?page-type=posts&post-id=${post.id}`
-        )
-      } else {
-        router.push(`/posts/${post.id}`)
-      }
-    } else {
-      router.push(`/posts/${post.id}`)
-    }
+    navigateToPost(post.id, spaceId)
   }
 
   return (
